Separate analog addon from page
