+++ www/src/Pages/AddonsConfigPage.jsx
@@ -22,6 +22,8 @@ import OnBoardLed, {
 	onBoardLedScheme,
 	onBoardLedState,
 } from '../Addons/OnBoardLed';
+import Analog, { analogScheme, analogState } from '../Addons/Analog';
+import AnalogPinOptions from '../Components/AnalogPinOptions';
 
 const I2C_BLOCKS = [
 	{ label: 'i2c0', value: 0 },
@@ -30,11 +32,6 @@ const I2C_BLOCKS = [
 
 const DUAL_STICK_MODES = [
 	{ label: 'D-Pad', value: 0 },
-	{ label: 'Left Analog', value: 1 },
-	{ label: 'Right Analog', value: 2 },
-];
-
-const ANALOG_STICK_MODES = [
 	{ label: 'Left Analog', value: 1 },
 	{ label: 'Right Analog', value: 2 },
 ];
@@ -56,15 +53,6 @@ const TILT_SOCD_MODES = [
 	{ label: 'Neutral', value: 1 },
 	{ label: 'Last Win', value: 2 },
 ];
-
-const INVERT_MODES = [
-	{ label: 'None', value: 0 },
-	{ label: 'X Axis', value: 1 },
-	{ label: 'Y Axis', value: 2 },
-	{ label: 'X/Y Axis', value: 3 },
-];
-
-const ANALOG_PINS = [26, 27, 28, 29];
 
 const TURBO_MASKS = [
 	{ label: 'None', value: 0 },
@@ -238,53 +226,6 @@ const schema = yup.object().shape({
 		.number()
 		.label('I2C Analog1219 Address')
 		.validateNumberWhenValue('I2CAnalog1219InputEnabled'),
-
-	AnalogInputEnabled: yup.number().required().label('Analog Input Enabled'),
-	analogAdc1PinX: yup
-		.number()
-		.label('Analog Stick 1 Pin X')
-		.validatePinWhenValue('AnalogInputEnabled'),
-	analogAdc1PinY: yup
-		.number()
-		.label('Analog Stick 1 Pin Y')
-		.validatePinWhenValue('AnalogInputEnabled'),
-	analogAdc1Mode: yup
-		.number()
-		.label('Analog Stick 1 Mode')
-		.validateSelectionWhenValue('AnalogInputEnabled', ANALOG_STICK_MODES),
-	analogAdc1Invert: yup
-		.number()
-		.label('Analog Stick 1 Invert')
-		.validateSelectionWhenValue('AnalogInputEnabled', INVERT_MODES),
-	analogAdc2PinX: yup
-		.number()
-		.label('Analog Stick 2 Pin X')
-		.validatePinWhenValue('AnalogInputEnabled'),
-	analogAdc2PinY: yup
-		.number()
-		.label('Analog Stick 2 Pin Y')
-		.validatePinWhenValue('AnalogInputEnabled'),
-	analogAdc2Mode: yup
-		.number()
-		.label('Analog Stick 2 Mode')
-		.validateSelectionWhenValue('AnalogInputEnabled', ANALOG_STICK_MODES),
-	analogAdc2Invert: yup
-		.number()
-		.label('Analog Stick 2 Invert')
-		.validateSelectionWhenValue('AnalogInputEnabled', INVERT_MODES),
-
-	forced_circularity: yup
-		.number()
-		.label('Force Circularity')
-		.validateRangeWhenValue('AnalogInputEnabled', 0, 1),
-	analog_deadzone: yup
-		.number()
-		.label('Deadzone Size (%)')
-		.validateRangeWhenValue('AnalogInputEnabled', 0, 100),
-	auto_calibrate: yup
-		.number()
-		.label('Auto Calibration')
-		.validateRangeWhenValue('AnalogInputEnabled', 0, 1),
 
 	FocusModeAddonEnabled: yup
 		.number()
@@ -612,6 +553,7 @@ const schema = yup.object().shape({
 		.number()
 		.label('WiiExtension I2C Speed')
 		.validateNumberWhenValue('WiiExtensionAddonEnabled'),
+	...analogScheme,
 	...bootselScheme,
 	...onBoardLedScheme,
 });
@@ -651,17 +593,7 @@ const defaultValues = {
 	tiltRightAnalogDownPin: -1,
 	tiltRightAnalogLeftPin: -1,
 	tiltRightAnalogRightPin: -1,
-	analogAdc1PinX: -1,
-	analogAdc1PinY: -1,
-	analogAdc1Mode: 1,
-	analogAdc1Invert: 0,
-	analogAdc2PinX: -1,
-	analogAdc2PinY: -1,
-	analogAdc2Mode: 2,
-	analogAdc2Invert: 0,
-	forced_circularity: 0,
-	analog_deadzone: 5,
-	auto_calibrate: 0,
+
 	buzzerPin: -1,
 	buzzerVolume: 100,
 	extrabuttonPin: -1,
@@ -695,7 +627,6 @@ const defaultValues = {
 	keyboardHostPinDplus: -1,
 	keyboardHostPin5V: -1,
 	keyboardHostMap: baseButtonMappings,
-	AnalogInputEnabled: 0,
 	FocusModeAddonEnabled: 0,
 	focusModeOledLockEnabled: 0,
 	focusModeRgbLockEnabled: 0,
@@ -713,6 +644,7 @@ const defaultValues = {
 	TurboInputEnabled: 0,
 	WiiExtensionAddonEnabled: 0,
 	SNESpadAddonEnabled: 0,
+	...analogState,
 	...bootselState,
 	...onBoardLedState,
 };
@@ -766,31 +698,11 @@ function flattenObject(object) {
 	return toReturn;
 }
 
-const AvailablePinOptions = ({ pins }) => {
-	const { t } = useTranslation();
-	return (
-		<>
-			<option value={-1}>
-				{t('AddonsConfig:analog-available-pins-option-not-set')}
-			</option>
-			{pins.map((i) => (
-				<option key={`analogPins-option-${i}`} value={i}>
-					{i}
-				</option>
-			))}
-		</>
-	);
-};
-
 export default function AddonsConfigPage() {
-	const { buttonLabels, updateUsedPins, usedPins } = useContext(AppContext);
+	const { buttonLabels, updateUsedPins } = useContext(AppContext);
 	const [saveMessage, setSaveMessage] = useState('');
 	const [storedData, setStoredData] = useState({});
 	const [validated, setValidated] = useState(false);
-
-	const availableAnalogPins = ANALOG_PINS.filter(
-		(pin) => !usedPins?.includes(pin),
-	);
 
 	const { t } = useTranslation();
 
@@ -862,195 +774,12 @@ export default function AddonsConfigPage() {
 						handleChange={handleChange}
 						handleCheckbox={handleCheckbox}
 					/>
-					<Section title={t('AddonsConfig:analog-header-text')}>
-						<div id="AnalogInputOptions" hidden={!values.AnalogInputEnabled}>
-							<p>{t('AddonsConfig:analog-warning')}</p>
-							<p>
-								{t('AddonsConfig:analog-available-pins-text', {
-									pins: availableAnalogPins.join(', '),
-								})}
-							</p>
-							<Row className="mb-3">
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-1-pin-x-label')}
-									name="analogAdc1PinX"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc1PinX}
-									error={errors.analogAdc1PinX}
-									isInvalid={errors.analogAdc1PinX}
-									onChange={handleChange}
-								>
-									<AvailablePinOptions pins={ANALOG_PINS} />
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-1-pin-y-label')}
-									name="analogAdc1PinY"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc1PinY}
-									error={errors.analogAdc1PinY}
-									isInvalid={errors.analogAdc1PinY}
-									onChange={handleChange}
-								>
-									<AvailablePinOptions pins={ANALOG_PINS} />
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-1-mode-label')}
-									name="analogAdc1Mode"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc1Mode}
-									error={errors.analogAdc1Mode}
-									isInvalid={errors.analogAdc1Mode}
-									onChange={handleChange}
-								>
-									{ANALOG_STICK_MODES.map((o, i) => (
-										<option
-											key={`button-analogAdc1Mode-option-${i}`}
-											value={o.value}
-										>
-											{o.label}
-										</option>
-									))}
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-1-invert-label')}
-									name="analogAdc1Invert"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc1Invert}
-									error={errors.analogAdc1Invert}
-									isInvalid={errors.analogAdc1Invert}
-									onChange={handleChange}
-								>
-									{INVERT_MODES.map((o, i) => (
-										<option
-											key={`button-analogAdc1Invert-option-${i}`}
-											value={o.value}
-										>
-											{o.label}
-										</option>
-									))}
-								</FormSelect>
-							</Row>
-							<Row className="mb-3">
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-2-pin-x-label')}
-									name="analogAdc2PinX"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc2PinX}
-									error={errors.analogAdc2PinX}
-									isInvalid={errors.analogAdc2PinX}
-									onChange={handleChange}
-								>
-									<AvailablePinOptions pins={ANALOG_PINS} />
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-2-pin-y-label')}
-									name="analogAdc2PinY"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc2PinY}
-									error={errors.analogAdc2PinY}
-									isInvalid={errors.analogAdc2PinY}
-									onChange={handleChange}
-								>
-									<AvailablePinOptions pins={ANALOG_PINS} />
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-2-mode-label')}
-									name="analogAdc2Mode"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc2Mode}
-									error={errors.analogAdc2Mode}
-									isInvalid={errors.analogAdc2Mode}
-									onChange={handleChange}
-								>
-									{ANALOG_STICK_MODES.map((o, i) => (
-										<option
-											key={`button-analogAdc2Mode-option-${i}`}
-											value={o.value}
-										>
-											{o.label}
-										</option>
-									))}
-								</FormSelect>
-								<FormSelect
-									label={t('AddonsConfig:analog-adc-2-invert-label')}
-									name="analogAdc2Invert"
-									className="form-select-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analogAdc2Invert}
-									error={errors.analogAdc2Invert}
-									isInvalid={errors.analogAdc2Invert}
-									onChange={handleChange}
-								>
-									{INVERT_MODES.map((o, i) => (
-										<option
-											key={`button-analogAdc2Invert-option-${i}`}
-											value={o.value}
-										>
-											{o.label}
-										</option>
-									))}
-								</FormSelect>
-							</Row>
-							<Row className="mb-3">
-								<FormControl
-									type="number"
-									label={t('AddonsConfig:analog-deadzone-size')}
-									name="analog_deadzone"
-									className="form-control-sm"
-									groupClassName="col-sm-3 mb-3"
-									value={values.analog_deadzone}
-									error={errors.analog_deadzone}
-									isInvalid={errors.analog_deadzone}
-									onChange={handleChange}
-									min={0}
-									max={100}
-								/>
-								<FormCheck
-									label={t('AddonsConfig:analog-force-circularity')}
-									type="switch"
-									id="Forced_circularity"
-									className="col-sm-3 ms-2"
-									isInvalid={false}
-									checked={Boolean(values.forced_circularity)}
-									onChange={(e) => {
-										handleCheckbox('forced_circularity', values);
-										handleChange(e);
-									}}
-								/>
-								<FormCheck
-									label={t('AddonsConfig:analog-auto-calibrate')}
-									type="switch"
-									id="Auto_calibrate"
-									className="col-sm-3 ms-2"
-									isInvalid={false}
-									checked={Boolean(values.auto_calibrate)}
-									onChange={(e) => {
-										handleCheckbox('auto_calibrate', values);
-										handleChange(e);
-									}}
-								/>
-							</Row>
-						</div>
-						<FormCheck
-							label={t('Common:switch-enabled')}
-							type="switch"
-							id="AnalogInputButton"
-							reverse
-							isInvalid={false}
-							checked={Boolean(values.AnalogInputEnabled)}
-							onChange={(e) => {
-								handleCheckbox('AnalogInputEnabled', values);
-								handleChange(e);
-							}}
-						/>
-					</Section>
+					<Analog
+						values={values}
+						errors={errors}
+						handleChange={handleChange}
+						handleCheckbox={handleCheckbox}
+					/>
 					<Section title={t('AddonsConfig:turbo-header-text')}>
 						<div id="TurboInputOptions" hidden={!values.TurboInputEnabled}>
 							<Row className="mb-3">
@@ -1088,7 +817,7 @@ export default function AddonsConfigPage() {
 									groupClassName="col-sm-3 mb-3"
 									value={values.turboShotCount}
 									error={errors.turboShotCount}
-									isInvalid={errors.turboShotCount}
+									isInvalid={Boolean(errors.turboShotCount)}
 									onChange={handleChange}
 									min={2}
 									max={30}
@@ -1103,7 +832,7 @@ export default function AddonsConfigPage() {
 									isInvalid={errors.pinShmupDial}
 									onChange={handleChange}
 								>
-									<AvailablePinOptions pins={ANALOG_PINS} />
+									<AnalogPinOptions />
 								</FormSelect>
 								<FormCheck
 									label={t('AddonsConfig:turbo-shmup-mode-label')}
